requirements.txt
aiogram==2.25.1
aiohttp==3.9.5
redis==5.0.7
python-dotenv==1.0.1
phonenumbers==8.13.45
